requirements.txt
airtable-python-wrapper==0.12.0
attrs==19.3.0
bottle==0.12.18
brother-ql==0.9.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
future==0.18.2
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
packbits==0.6
Pillow==7.0.0
pyusb==1.0.2
requests==2.23.0
six==1.14.0
urllib3==1.25.8
